test: dto와 관련된 테스트를 dto 테스트 클래스로 이동
