+++ backend/ddang/src/test/java/com/ddang/ddang/auction/application/AuctionServiceTest.java
@@ -197,15 +197,6 @@ class AuctionServiceTest extends AuctionServiceFixture {
     }
 
     @Test
-    void 지정한_아이디에_해당하는_경매의_판매자_신뢰도가_null이라면_서비스에서_반환하는_dto에서_판매자_신뢰도를_나타내는_부분도_null이다() {
-        // when
-        final ReadAuctionDto actual = auctionService.readByAuctionId(신뢰도가_null인_판매자의_경매.getId());
-
-        // then
-        assertThat(actual.sellerReliability()).isNull();
-    }
-
-    @Test
     void 첫번째_페이지의_경매_목록을_조회한다() {
         // when
         final ReadAuctionsDto actual = auctionService.readAllByCondition(
